entropy/requirements.txt
aioredis==1.3.1
amqp==2.6.0
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.5
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.7
djangorestframework==3.11.0
future==0.18.2
h11==0.8.1
h2==3.2.0
hiredis==1.0.1
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
kombu==4.6.10
more-itertools==8.4.0
msgpack==0.6.2
packaging==20.4
pluggy==0.13.1
py==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==5.4.3
pytest-asyncio==0.14.0
pytest-django==3.9.0
pytz==2020.1
redis==3.5.3
rfc3986==1.4.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.5
zope.interface==5.1.0
